requirements.txt
# Knowledge Graph Management Workbench - Python backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pandas>=2.0.0
openpyxl>=3.1.0
google-genai>=1.0.0
